save features of each run.
start of SNR (signal to noise ratio)

We want to train 64 backbones from scratch and use the best one.

Here in each run the features are saved in a file features/feat{id_run}{1 or 5} with 1 or 5 for 1-shot or 5-shot classify.
I had to add the i_run variable to let update_few_shot_meta_data function know in which run it is.